Definitions and registry changes
Processes are now defined via define-process and must be spawned via spawn
(or lower-level do-spawn).

Processes are no longer "named" but instead get a numeric ID on spawn.
Multiple processes of the same type (as defined by define-process) can
be spawned at once, each its own instance with its own ID. This is a
useful abstraction.

Processes can be registered to a name via (register name process) which
allows processes to be looked up by name instead of just ID. This is not
only useful for referncing locally, but will be essential once processes
can communicate over the network.